Model#find returns nil for a nil id
To be more precise: It returs nil for an id with an empty string
representation.

Reason: Appending (almost) nothing to the class' url will result in
an url almost always specifying the same (class "index") resource and
not the required instance resource. And this will not return a 404
(which is converted to nil), but a successful index request.